Move `controllers` to `internal/controllers`
There is no good reason for it to be exposed and available through a
public API, and this follows the new kubebuilder defaults.
