+++ Dockerfile
@@ -25,7 +25,6 @@ RUN go mod download
 
 # copy source code
 COPY main.go main.go
-COPY controllers/ controllers/
 COPY internal/ internal/
 
 # build without specifing the arch
